+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@antv/util",
-  "version": "3.3.4",
+  "version": "3.3.5",
   "license": "MIT",
   "sideEffects": false,
   "main": "lib/index.js",
