+++ src/dom/index.ts
@@ -1,0 +1,2 @@
+export { createDOM } from './create-dom';
+export { modifyCSS } from './modify-css';
