+++ src/index.ts
@@ -3,3 +3,4 @@ export * from './matrix';
 export * from './path';
 export * from './lodash';
 export * from './math';
+export * from  './dom';
