+++ src/matrix/vertical.ts
@@ -1,0 +1,17 @@
+/**
+ * 计算二维向量的垂直向量
+ * @param out
+ * @param v
+ * @param flag
+ */
+export function vertical(out: number[], v: number[], flag: boolean): number[] {
+  if (flag) {
+    out[0] = v[1];
+    out[1] = -1 * v[0];
+  } else {
+    out[0] = -1 * v[1];
+    out[1] = v[0];
+  }
+
+  return out;
+}
